refactor: deduplicate slug field
