+++ apps/historia/src/collections/Categories.ts
@@ -2,7 +2,7 @@ import type { CollectionConfig } from 'payload';
 
 import { anyone } from '../access/anyone';
 import { authenticated } from '../access/authenticated';
-import { slugField } from '@/slug';
+import { slugField } from '@/fields/slug';
 
 export const Categories: CollectionConfig = {
   slug: 'topics',
