--- apps/historia/src/slug/formatSlug.ts
@@ -1,25 +1,0 @@
-import type { FieldHook } from 'payload'
-
-export const formatSlug = (val: string): string =>
-  val
-    .replace(/ /g, '-')
-    .replace(/[^\w-]+/g, '')
-    .toLowerCase()
-
-export const formatSlugHook =
-  (fallback: string): FieldHook =>
-  ({ data, operation, originalDoc, value }) => {
-    if (typeof value === 'string') {
-      return formatSlug(value)
-    }
-
-    if (operation === 'create' || !data?.slug) {
-      const fallbackData = data?.[fallback] || data?.[fallback]
-
-      if (fallbackData && typeof fallbackData === 'string') {
-        return formatSlug(fallbackData)
-      }
-    }
-
-    return value
-  }
