--- apps/historia/src/slug/index.scss
@@ -1,12 +1,0 @@
-.slug-field-component {
-  .label-wrapper {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-
-  .lock-button {
-    margin: 0;
-    padding-bottom: 0.3125rem;
-  }
-}
